backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
sentence-transformers==3.3.1
chromadb==0.5.23
google-generativeai==0.8.3
firecrawl-py
python-dotenv==1.0.1
pydantic==2.10.3
lxml==5.3.0
